requirements.txt
joblib
numpy
pandas
Pillow
PyYAML
scikit-learn
seaborn
torch
torchvision
tqdm
matplotlib
